Added interface and changed file names
