+++ src/main/java/com/sparta/ps/kimchi/DAOEnabler.java
@@ -1,0 +1,2 @@
+package com.sparta.ps.kimchi;public class DAOEnabler {
+}
